requirements.txt
python-telegram-bot[job-queue]==20.7
apscheduler==3.10.4
pytz

fastapi
uvicorn